ProxyChecker
# Proxy Checker

This program tests a list of proxies and saves the working ones to a file. It is designed to help you check if your proxies are active and functional.

## Developer

- Developer Zehir
- GitHub [httpsgithub.comZehirnoname](httpsgithub.comZehirnoname)

---

## How to Download and Run the Program

Follow these steps to download and run the program on your computer.

### 1. Clone the GitHub Repository

To download the project, you need to have Git installed on your machine. You can install Git from [here](httpsgit-scm.comdownloads).

Follow these steps to clone the repository

1. Go to the GitHub repository [httpsgithub.comZehirnonameProxyChecker](httpsgithub.comZehirnonameProxyChecker).
2. Click on the green Code button and copy the HTTPS link.

    Example
    ```
    httpsgithub.comZehirnonameProxyChecker.git
    ```

3. Open a terminal (command prompt) and type the following command to clone the repository

    ```bash
    git clone httpsgithub.comZehirnonameProxyChecker.git
    ```

4. Navigate into the project directory

    ```bash
    cd ProxyChecker
    ```

### 2. Install Python and Required Libraries

The program is written in Python. Make sure you have Python installed on your system.

#### Check if Python is installed

Open a terminal or command prompt and type

```bash
python --version